Refactor object detection
